examples/BraTS/metrics/project/requirements.txt
PyYAML~=5.3
pandas
typer
MedPy
SimpleITK
numpy
scikit-learn
synapseclient
nibabel